fix: segfault on parameter parsing

+++ src/tsp_instance.c
@@ -58,7 +58,7 @@ void fill_edge_cost_matrix(const TspInstance* instance)
 
 TspInstance* initialize_random_tsp_instance(const TspParams* params)
 {
-    const long number_of_nodes = params->number_of_nodes;
+    const unsigned int number_of_nodes = params->number_of_nodes;
     srand(params->seed);
     Node* nodes = malloc(number_of_nodes * sizeof(Node));
     if (!nodes)
@@ -68,6 +68,7 @@ TspInstance* initialize_random_tsp_instance(const TspParams* params)
     }
     for (int i = 0; i < number_of_nodes; i++)
     {
+        //TODO Con questo metodo i nodi non possono avere la virgola pensare di passare a nodi non interi
         nodes[i].x = params->generation_area.x_square + rand() % (params->generation_area.square_side + 1);
         nodes[i].y = params->generation_area.y_square + rand() % (params->generation_area.square_side + 1);
     }
